requirements.txt
flask
Pillow
gunicorn
rembg
onnxruntime
